update push notification event model
