+++ sdk/communication/AzureCommunicationChat/Tests/Recordings/test_ListThreads_ReturnsChatThreadItems.json
@@ -1,96 +1,86 @@
 {
-  "name" : "test_ListThreads_ReturnsChatThreadItems",
   "interactions" : [
     {
       "response" : {
-        "url" : "https:\/\/acs-gnygqibv4wh32.communication.azure.com\/chat\/threads?api-version=2021-09-07",
-        "headers" : {
-          "x-azure-ref" : "0JPvRYgAAAAA1QrYJdcIfTrpXMSqjMqgXV1NURURHRTA4MTkAOWZjN2I1MTktYThjYy00Zjg5LTkzNWUtYzkxNDhhZTA5ZTgx",
-          "ms-cv" : "9s7u+ulhgEiPtCJBpVF0ig.0",
-          "Location" : "https:\/\/acs-gnygqibv4wh32.communication.azure.com\/chat\/threads\/19%3A9x7TT_Df3uFQjQqSRC-Oj0s8mpcq-gyGpjFflvT6dPQ1@thread.v2",
-          "api-supported-versions" : "2021-03-07, 2021-04-05-preview6, 2021-09-07, 2022-11-15-preview8",
-          "x-processing-time" : "467ms",
-          "Strict-Transport-Security" : "max-age=2592000",
-          "x-cache" : "CONFIG_NOCACHE",
-          "Content-Type" : "application\/json; charset=utf-8"
-        },
         "status" : 201,
         "body" : {
           "chatThread" : {
-            "id" : "19:9x7TT_Df3uFQjQqSRC-Oj0s8mpcq-gyGpjFflvT6dPQ1@thread.v2",
-            "createdOn" : "2022-07-15T23:41:24Z",
+            "id" : "19:SigYcbplgT3ZObJBxHDTeu6H_ulbF_brYuDNnpjA2P81@thread.v2",
+            "createdOn" : "2022-08-16T00:06:53Z",
             "topic" : "Test list threads",
             "createdByCommunicationIdentifier" : {
-              "rawId" : "8:acs:109dd97f-0bec-496a-9faa-fb9fe86fd940_00000012-a417-0ffa-0d8b-08482200801b",
+              "rawId" : "8:acs:d19c0cec-7b5e-47c8-842d-bfbd3c8bd7db_00000013-43d9-f9db-78fe-343a0d00b30f",
               "communicationUser" : {
-                "id" : "8:acs:109dd97f-0bec-496a-9faa-fb9fe86fd940_00000012-a417-0ffa-0d8b-08482200801b"
+                "id" : "8:acs:d19c0cec-7b5e-47c8-842d-bfbd3c8bd7db_00000013-43d9-f9db-78fe-343a0d00b30f"
               }
             }
           }
+        },
+        "url" : "https:\/\/acs-gnygqibv4wh32.communication.azure.com\/chat\/threads?api-version=2021-09-07",
+        "headers" : {
+          "x-azure-ref" : "0nd\/6YgAAAAA1JCxXG6EjT6ArXKJVUJNrWVZSMzExMDAwMTE2MDE3ADlmYzdiNTE5LWE4Y2MtNGY4OS05MzVlLWM5MTQ4YWUwOWU4MQ==",
+          "Content-Type" : "application\/json; charset=utf-8",
+          "x-cache" : "CONFIG_NOCACHE",
+          "Strict-Transport-Security" : "max-age=2592000",
+          "api-supported-versions" : "2021-03-07, 2021-04-05-preview6, 2021-09-07, 2022-11-15-preview8",
+          "Location" : "https:\/\/acs-gnygqibv4wh32.communication.azure.com\/chat\/threads\/19%3ASigYcbplgT3ZObJBxHDTeu6H_ulbF_brYuDNnpjA2P81@thread.v2",
+          "ms-cv" : "C3VzIYPo+Ee4C3mzM8178A.0",
+          "x-processing-time" : "803ms"
         }
       },
+      "recorded_at" : 1660608414.281776,
       "request" : {
-        "url" : "https:\/\/endpoint\/chat\/threads?",
         "body" : {
           "topic" : "Test list threads"
         },
-        "headers" : {
-          "User-Agent" : "azsdk-ios-AzureCommunicationChatTests\/1.0 (MacBookPro16,1 - 21.5.0; xctest:13.3 -> iOS 9.0; en_US)",
-          "repeatability-request-id" : "A1CE4CD8-7FC7-49F2-A3FB-AD68C5B9762F",
-          "Content-Type" : "application\/json",
-          "Accept" : "application\/json"
-        },
-        "method" : "POST"
-      },
-      "recorded_at" : 1657928484.8478489
-    },
-    {
-      "request" : {
-        "method" : "GET",
+        "method" : "POST",
         "url" : "https:\/\/endpoint\/chat\/threads?",
         "headers" : {
-          "User-Agent" : "azsdk-ios-AzureCommunicationChatTests\/1.0 (MacBookPro16,1 - 21.5.0; xctest:13.3 -> iOS 9.0; en_US)",
+          "Content-Type" : "application\/json",
+          "User-Agent" : "azsdk-ios-AzureCommunicationChatTests\/1.0 (MacBookPro16,1 - 21.6.0; xctest:13.3 -> iOS 9.0; en_US)",
+          "repeatability-request-id" : "1BAF3AA4-15C7-4DEB-8AF5-0617B1009B48",
           "Accept" : "application\/json"
         }
-      },
-      "recorded_at" : 1657928484.9791799,
+      }
+    },
+    {
+      "recorded_at" : 1660608414.477138,
       "response" : {
-        "headers" : {
-          "x-processing-time" : "100ms",
-          "x-azure-ref" : "0JPvRYgAAAADh0\/7tVqv8QZ476eFT8CztV1NURURHRTA4MTkAOWZjN2I1MTktYThjYy00Zjg5LTkzNWUtYzkxNDhhZTA5ZTgx",
-          "api-supported-versions" : "2021-03-07, 2021-04-05-preview6, 2021-09-07, 2022-11-15-preview8",
-          "ms-cv" : "CuzDnJk1R0aIKXMf7+bdug.0",
-          "x-cache" : "CONFIG_NOCACHE",
-          "Strict-Transport-Security" : "max-age=2592000",
-          "Content-Type" : "application\/json; charset=utf-8"
-        },
-        "status" : 200,
+        "url" : "https:\/\/acs-gnygqibv4wh32.communication.azure.com\/chat\/threads?api-version=2021-09-07",
         "body" : {
           "value" : [
             {
-              "id" : "19:9x7TT_Df3uFQjQqSRC-Oj0s8mpcq-gyGpjFflvT6dPQ1@thread.v2",
+              "id" : "19:SigYcbplgT3ZObJBxHDTeu6H_ulbF_brYuDNnpjA2P81@thread.v2",
               "topic" : "Test list threads",
-              "lastMessageReceivedOn" : "2022-07-15T23:41:24Z"
+              "lastMessageReceivedOn" : "2022-08-16T00:06:53Z"
             },
             {
-              "id" : "19:yKQMg02kdzTIWCIcfpq6SvXYAMDbSBjNPbMkjJmEgI01@thread.v2",
+              "id" : "19:7jUrhTw8S6t31erkWAjBUFuBhyLWGUxaYHunLGj3ODI1@thread.v2",
               "topic" : "Test topic",
-              "lastMessageReceivedOn" : "2022-07-15T23:41:24Z"
-            },
-            {
-              "id" : "19:l1Advf7jnmGuBIDFz9dkbsij-5dBfr52Z3-UK_zRXZg1@thread.v2",
-              "topic" : "Test list threads",
-              "lastMessageReceivedOn" : "2022-07-15T23:39:15Z"
-            },
-            {
-              "id" : "19:V1eQc-26k7LMb1Es8sHkq-4H1BfVNXD20GyVi5TSUIs1@thread.v2",
-              "topic" : "Test topic",
-              "lastMessageReceivedOn" : "2022-07-15T23:39:14Z"
+              "lastMessageReceivedOn" : "2022-08-16T00:06:53Z"
             }
           ]
         },
-        "url" : "https:\/\/acs-gnygqibv4wh32.communication.azure.com\/chat\/threads?api-version=2021-09-07"
+        "status" : 200,
+        "headers" : {
+          "x-azure-ref" : "0nt\/6YgAAAACpoQZZSq79QKaqijEUTPJuWVZSMzExMDAwMTE2MDE3ADlmYzdiNTE5LWE4Y2MtNGY4OS05MzVlLWM5MTQ4YWUwOWU4MQ==",
+          "x-cache" : "CONFIG_NOCACHE",
+          "x-processing-time" : "118ms",
+          "Strict-Transport-Security" : "max-age=2592000",
+          "Content-Type" : "application\/json; charset=utf-8",
+          "ms-cv" : "rV7can5Qakuj9MeTgK7hPg.0",
+          "api-supported-versions" : "2021-03-07, 2021-04-05-preview6, 2021-09-07, 2022-11-15-preview8"
+        }
+      },
+      "request" : {
+        "url" : "https:\/\/endpoint\/chat\/threads?",
+        "method" : "GET",
+        "headers" : {
+          "Accept" : "application\/json",
+          "User-Agent" : "azsdk-ios-AzureCommunicationChatTests\/1.0 (MacBookPro16,1 - 21.6.0; xctest:13.3 -> iOS 9.0; en_US)"
+        }
       }
     }
-  ]
+  ],
+  "name" : "test_ListThreads_ReturnsChatThreadItems"
 }
