+++ sdk/communication/AzureCommunicationChat/Tests/Recordings/test_StartPushNotifications_ReturnsSuccess.json
@@ -1,22 +1,13 @@
 {
-  "name" : "test_StartPushNotifications_ReturnsSuccess",
   "interactions" : [
     {
-      "response" : {
-        "url" : "https:\/\/edge.skype.com\/registrar\/prod\/v2\/registrations",
-        "headers" : {
-          "x-cache" : "CONFIG_NOCACHE",
-          "ms-cv" : "pKRoR+QiP0qmCCVv\/0V27g.0",
-          "x-content-type-options" : "nosniff",
-          "x-msedge-ref" : "Ref A: 1B7F261EE3D04C168CD637BE2E6CFB5D Ref B: WSTEDGE0710 Ref C: 2022-07-15T23:41:25Z",
-          "Content-Length" : "0",
-          "Strict-Transport-Security" : "max-age=31536000; includeSubDomains"
-        },
-        "status" : 202,
-        "body" : ""
-      },
-      "recorded_at" : 1657928485.2480569,
       "request" : {
+        "method" : "POST",
+        "headers" : {
+          "Content-Type" : "application\/json",
+          "User-Agent" : "azsdk-ios-AzureCommunicationChatTests\/1.0 (MacBookPro16,1 - 21.6.0; xctest:13.3 -> iOS 9.0; en_US)"
+        },
+        "url" : "https:\/\/edge.skype.com\/registrar\/prod\/v2\/registrations",
         "body" : {
           "clientDescription" : {
             "templateKey" : "AcsIos.AcsNotify_Chat_3.0",
@@ -39,14 +30,23 @@
           },
           "registrationId" : "0E0F0BE0-0000-00C0-B000-A00A00E00BD0",
           "nodeId" : ""
-        },
-        "method" : "POST",
-        "url" : "https:\/\/edge.skype.com\/registrar\/prod\/v2\/registrations",
-        "headers" : {
-          "User-Agent" : "azsdk-ios-AzureCommunicationChatTests\/1.0 (MacBookPro16,1 - 21.5.0; xctest:13.3 -> iOS 9.0; en_US)",
-          "Content-Type" : "application\/json"
         }
-      }
+      },
+      "response" : {
+        "headers" : {
+          "Content-Length" : "0",
+          "Strict-Transport-Security" : "max-age=31536000; includeSubDomains",
+          "x-msedge-ref" : "Ref A: 653CC3DF62C0464C9DF5B90C0B4D036E Ref B: YVR311000111033 Ref C: 2022-08-16T00:06:54Z",
+          "x-cache" : "CONFIG_NOCACHE",
+          "ms-cv" : "4CI+75DBi0OB8N420uOPoA.0",
+          "x-content-type-options" : "nosniff"
+        },
+        "body" : "",
+        "url" : "https:\/\/edge.skype.com\/registrar\/prod\/v2\/registrations",
+        "status" : 202
+      },
+      "recorded_at" : 1660608414.8790212
     }
-  ]
+  ],
+  "name" : "test_StartPushNotifications_ReturnsSuccess"
 }
