+++ sdk/communication/AzureCommunicationChat/Tests/Recordings/test_StopPushNotifications_ReturnsSuccess.json
@@ -2,27 +2,26 @@
   "name" : "test_StopPushNotifications_ReturnsSuccess",
   "interactions" : [
     {
+      "recorded_at" : 1660608415.139955,
       "response" : {
         "url" : "https:\/\/edge.skype.com\/registrar\/prod\/v2\/registrations",
         "status" : 202,
         "headers" : {
           "x-content-type-options" : "nosniff",
-          "x-msedge-ref" : "Ref A: 065DB4A7181D453AABADA0B17699F07D Ref B: WSTEDGE0710 Ref C: 2022-07-15T23:41:25Z",
+          "Strict-Transport-Security" : "max-age=31536000; includeSubDomains",
+          "x-msedge-ref" : "Ref A: F27FC3994FF84A4B82A8B5E823F18D01 Ref B: YVR311000111033 Ref C: 2022-08-16T00:06:54Z",
           "x-cache" : "CONFIG_NOCACHE",
           "Content-Length" : "0",
-          "Strict-Transport-Security" : "max-age=31536000; includeSubDomains",
-          "ms-cv" : "O28j1rpi3UWuHfoaf4Wdvw.0"
+          "ms-cv" : "VKmJRd1kY06qJc+GGySUeA.0"
         },
         "body" : ""
       },
-      "recorded_at" : 1657928485.5425019,
       "request" : {
         "method" : "POST",
         "headers" : {
           "Content-Type" : "application\/json",
-          "User-Agent" : "azsdk-ios-AzureCommunicationChatTests\/1.0 (MacBookPro16,1 - 21.5.0; xctest:13.3 -> iOS 9.0; en_US)"
+          "User-Agent" : "azsdk-ios-AzureCommunicationChatTests\/1.0 (MacBookPro16,1 - 21.6.0; xctest:13.3 -> iOS 9.0; en_US)"
         },
-        "url" : "https:\/\/edge.skype.com\/registrar\/prod\/v2\/registrations",
         "body" : {
           "clientDescription" : {
             "templateKey" : "AcsIos.AcsNotify_Chat_3.0",
@@ -45,31 +44,32 @@
           },
           "registrationId" : "0E0F0BE0-0000-00C0-B000-A00A00E00BD0",
           "nodeId" : ""
-        }
+        },
+        "url" : "https:\/\/edge.skype.com\/registrar\/prod\/v2\/registrations"
       }
     },
     {
-      "response" : {
-        "url" : "https:\/\/edge.skype.com\/registrar\/prod\/v2\/registrations\/0E0F0BE0-0000-00C0-B000-A00A00E00BD0",
-        "body" : "",
-        "headers" : {
-          "x-cache" : "CONFIG_NOCACHE",
-          "ms-cv" : "LseFs9v7Xkm7RNiHkdR6gQ.0",
-          "Content-Length" : "0",
-          "x-msedge-ref" : "Ref A: E07FF53DB488419D9B547649F03D6455 Ref B: WSTEDGE0710 Ref C: 2022-07-15T23:41:25Z",
-          "x-content-type-options" : "nosniff",
-          "Strict-Transport-Security" : "max-age=31536000; includeSubDomains"
-        },
-        "status" : 202
-      },
-      "recorded_at" : 1657928485.8068218,
       "request" : {
         "url" : "https:\/\/edge.skype.com\/registrar\/prod\/v2\/registrations\/0E0F0BE0-0000-00C0-B000-A00A00E00BD0",
         "headers" : {
-          "Content-Type" : "application\/json",
-          "User-Agent" : "azsdk-ios-AzureCommunicationChatTests\/1.0 (MacBookPro16,1 - 21.5.0; xctest:13.3 -> iOS 9.0; en_US)"
+          "User-Agent" : "azsdk-ios-AzureCommunicationChatTests\/1.0 (MacBookPro16,1 - 21.6.0; xctest:13.3 -> iOS 9.0; en_US)",
+          "Content-Type" : "application\/json"
         },
         "method" : "DELETE"
+      },
+      "recorded_at" : 1660608415.454143,
+      "response" : {
+        "headers" : {
+          "x-msedge-ref" : "Ref A: EBDE2226E08642548401895D755B3E53 Ref B: YVR311000111033 Ref C: 2022-08-16T00:06:55Z",
+          "x-cache" : "CONFIG_NOCACHE",
+          "Strict-Transport-Security" : "max-age=31536000; includeSubDomains",
+          "x-content-type-options" : "nosniff",
+          "ms-cv" : "RKFx8HVusEe1UJEGQpwCmA.0",
+          "Content-Length" : "0"
+        },
+        "url" : "https:\/\/edge.skype.com\/registrar\/prod\/v2\/registrations\/0E0F0BE0-0000-00C0-B000-A00A00E00BD0",
+        "body" : "",
+        "status" : 202
       }
     }
   ]
