+++ sdk/communication/AzureCommunicationChat/Tests/Recordings/test_deleteRegistration_ReturnsSuccess.json
@@ -1,8 +1,26 @@
 {
+  "name" : "test_deleteRegistration_ReturnsSuccess",
   "interactions" : [
     {
-      "recorded_at" : 1657928460.44611,
+      "response" : {
+        "status" : 202,
+        "body" : "",
+        "url" : "https:\/\/edge.skype.com\/registrar\/prod\/v2\/registrations",
+        "headers" : {
+          "x-cache" : "CONFIG_NOCACHE",
+          "Strict-Transport-Security" : "max-age=31536000; includeSubDomains",
+          "x-msedge-ref" : "Ref A: 910026674BA246349B757D92C01ADC26 Ref B: YVR311000108023 Ref C: 2022-08-16T00:07:23Z",
+          "ms-cv" : "pTFcxGhx+0+50xqq8Bg2ww.0",
+          "x-content-type-options" : "nosniff",
+          "Content-Length" : "0"
+        }
+      },
       "request" : {
+        "headers" : {
+          "Content-Type" : "application\/json",
+          "User-Agent" : "azsdk-ios-AzureCommunicationChatTests\/1.0 (MacBookPro16,1 - 21.6.0; xctest:13.3 -> iOS 9.0; en_US)"
+        },
+        "method" : "POST",
         "body" : {
           "clientDescription" : {
             "templateKey" : "AcsIos.AcsNotify_Chat_3.0",
@@ -26,51 +44,33 @@
           "registrationId" : "0E0F0BE0-0000-00C0-B000-A00A00E00BD0",
           "nodeId" : ""
         },
-        "method" : "POST",
-        "url" : "https:\/\/edge.skype.com\/registrar\/prod\/v2\/registrations",
-        "headers" : {
-          "User-Agent" : "azsdk-ios-AzureCommunicationChatTests\/1.0 (MacBookPro16,1 - 21.5.0; xctest:13.3 -> iOS 9.0; en_US)",
-          "Content-Type" : "application\/json"
-        }
+        "url" : "https:\/\/edge.skype.com\/registrar\/prod\/v2\/registrations"
       },
-      "response" : {
-        "body" : "",
-        "url" : "https:\/\/edge.skype.com\/registrar\/prod\/v2\/registrations",
-        "status" : 202,
-        "headers" : {
-          "x-content-type-options" : "nosniff",
-          "Strict-Transport-Security" : "max-age=31536000; includeSubDomains",
-          "x-cache" : "CONFIG_NOCACHE",
-          "x-msedge-ref" : "Ref A: 090AF5F0B071491CB5C7D9C1718C6677 Ref B: WSTEDGE0920 Ref C: 2022-07-15T23:41:00Z",
-          "Content-Length" : "0",
-          "ms-cv" : "s8MW04\/kYkSd1TBD+zPSHg.0"
-        }
-      }
+      "recorded_at" : 1660608443.9042571
     },
     {
-      "response" : {
-        "status" : 202,
-        "body" : "",
+      "request" : {
         "url" : "https:\/\/edge.skype.com\/registrar\/prod\/v2\/registrations\/0E0F0BE0-0000-00C0-B000-A00A00E00BD0",
         "headers" : {
-          "ms-cv" : "hUIlkO+wiU2b+jWYeaJLAQ.0",
-          "x-msedge-ref" : "Ref A: 52739CD5199C486EB2D4B82C7CD2CDEC Ref B: WSTEDGE0920 Ref C: 2022-07-15T23:41:00Z",
-          "Strict-Transport-Security" : "max-age=31536000; includeSubDomains",
+          "Content-Type" : "application\/json",
+          "User-Agent" : "azsdk-ios-AzureCommunicationChatTests\/1.0 (MacBookPro16,1 - 21.6.0; xctest:13.3 -> iOS 9.0; en_US)"
+        },
+        "method" : "DELETE"
+      },
+      "response" : {
+        "url" : "https:\/\/edge.skype.com\/registrar\/prod\/v2\/registrations\/0E0F0BE0-0000-00C0-B000-A00A00E00BD0",
+        "status" : 202,
+        "body" : "",
+        "headers" : {
           "x-content-type-options" : "nosniff",
           "x-cache" : "CONFIG_NOCACHE",
+          "ms-cv" : "sSeo2ZzzNkiRGP243VA4nQ.0",
+          "x-msedge-ref" : "Ref A: 1077F1100B8E4FFF864D560DAF8470DB Ref B: YVR311000108023 Ref C: 2022-08-16T00:07:23Z",
+          "Strict-Transport-Security" : "max-age=31536000; includeSubDomains",
           "Content-Length" : "0"
         }
       },
-      "request" : {
-        "headers" : {
-          "User-Agent" : "azsdk-ios-AzureCommunicationChatTests\/1.0 (MacBookPro16,1 - 21.5.0; xctest:13.3 -> iOS 9.0; en_US)",
-          "Content-Type" : "application\/json"
-        },
-        "method" : "DELETE",
-        "url" : "https:\/\/edge.skype.com\/registrar\/prod\/v2\/registrations\/0E0F0BE0-0000-00C0-B000-A00A00E00BD0"
-      },
-      "recorded_at" : 1657928460.6515231
+      "recorded_at" : 1660608444.146955
     }
-  ],
-  "name" : "test_deleteRegistration_ReturnsSuccess"
+  ]
 }
