+++ sdk/communication/AzureCommunicationChat/Tests/Recordings/test_setRegistration_ReturnsSuccess.json
@@ -2,12 +2,12 @@
   "name" : "test_setRegistration_ReturnsSuccess",
   "interactions" : [
     {
+      "recorded_at" : 1660608444.416357,
       "request" : {
         "method" : "POST",
-        "url" : "https:\/\/edge.skype.com\/registrar\/prod\/v2\/registrations",
         "headers" : {
-          "User-Agent" : "azsdk-ios-AzureCommunicationChatTests\/1.0 (MacBookPro16,1 - 21.5.0; xctest:13.3 -> iOS 9.0; en_US)",
-          "Content-Type" : "application\/json"
+          "Content-Type" : "application\/json",
+          "User-Agent" : "azsdk-ios-AzureCommunicationChatTests\/1.0 (MacBookPro16,1 - 21.6.0; xctest:13.3 -> iOS 9.0; en_US)"
         },
         "body" : {
           "clientDescription" : {
@@ -31,21 +31,21 @@
           },
           "registrationId" : "0E0F0BE0-0000-00C0-B000-A00A00E00BD0",
           "nodeId" : ""
-        }
+        },
+        "url" : "https:\/\/edge.skype.com\/registrar\/prod\/v2\/registrations"
       },
-      "recorded_at" : 1657928305.0803099,
       "response" : {
-        "url" : "https:\/\/edge.skype.com\/registrar\/prod\/v2\/registrations",
         "headers" : {
+          "x-msedge-ref" : "Ref A: 1FCF1FE7388A4E49BBDC992849DED969 Ref B: YVR311000108023 Ref C: 2022-08-16T00:07:24Z",
+          "x-cache" : "CONFIG_NOCACHE",
           "Strict-Transport-Security" : "max-age=31536000; includeSubDomains",
-          "x-content-type-options" : "nosniff",
-          "ms-cv" : "oLOnXbtgJEeZ+qbpwKEG8g.0",
+          "ms-cv" : "w1LJmZj9YES1qSgl9nb1\/g.0",
           "Content-Length" : "0",
-          "x-msedge-ref" : "Ref A: BCDB8CF23FB64A3B8B40FED1FD71CF32 Ref B: WSTEDGE0608 Ref C: 2022-07-15T23:38:24Z",
-          "x-cache" : "CONFIG_NOCACHE"
+          "x-content-type-options" : "nosniff"
         },
         "status" : 202,
-        "body" : ""
+        "body" : "",
+        "url" : "https:\/\/edge.skype.com\/registrar\/prod\/v2\/registrations"
       }
     }
   ]
